BugFix [0000429]:
It's possible to disable all UI plugins, now the plugin loader tries to obey the user preference for clist, but if all fails, it will try loading a clist ignoring the ignore information.

It's also now possible to deselect all UI plugins from the options, however the plugin loader will ignore that info if it has to, before showing the error box.